Create javascript-millisecond-text-input-update.js
This javascript file provides the js_millisecond_update() function which can be used for ping tests with the executium API version 2. This script can also be adapted should a user wish to update a input value with a millisecond output. The millisecond javascript function is all provided in vanilla javascript.